Add support for grabbing TTY input
termbox reads data from the tty and generates events for those. If an
application wants to temporary take over the input stream, it can call
GrabTtyInput() which returns an io.ReadCloser which returns bytes read from the
input stream without generating events.

This can be useful if launching another application which wants to read from the
tty.