Input grabber: Fix goroutine deadlock
When the TTY input was grabbed, the `sigio` case block in the select sent a
message on the `inpgrab` channel, which in turn sent data to the `inpgrab_ch`
channel for the application to read from. The `inpgrab` case block is in the
same select block as the `sigio` case block, and this caused a deadlock.

This was because when input data came in too fast, it was possible for the
`inpgrab` channel's buffer to be full, making the channel send operation from
the `sigio` block to be blocked. Since the operation to read from the `inpgrab`
channel was in the same select block, it never got executed.

The fix removes the deadlock by directly sending input data to the `inpgrab_ch`
channel in the `sigio` case block. That channel gets drained whenever the
application reads from it.